ref: Move utils, consts out of integrations/opentelemetry/
