+++ sentry_sdk/opentelemetry/span_processor.py
@@ -15,14 +15,12 @@ import sentry_sdk
 from sentry_sdk.consts import SPANDATA
 from sentry_sdk.tracing import DEFAULT_SPAN_ORIGIN
 from sentry_sdk.utils import get_current_thread_meta
-from sentry_sdk.opentelemetry.sampler import create_sampling_context
-from sentry_sdk.profiler.continuous_profiler import (
-    try_autostart_continuous_profiler,
-    get_profiler_id,
-    try_profile_lifecycle_trace_start,
+from sentry_sdk.opentelemetry.consts import (
+    OTEL_SENTRY_CONTEXT,
+    SentrySpanAttribute,
 )
-from sentry_sdk.profiler.transaction_profiler import Profile
-from sentry_sdk.integrations.opentelemetry.utils import (
+from sentry_sdk.opentelemetry.sampler import create_sampling_context
+from sentry_sdk.opentelemetry.utils import (
     is_sentry_span,
     convert_from_otel_timestamp,
     extract_span_attributes,
@@ -33,10 +31,12 @@ from sentry_sdk.integrations.opentelemetry.utils import (
     get_sentry_meta,
     set_sentry_meta,
 )
-from sentry_sdk.integrations.opentelemetry.consts import (
-    OTEL_SENTRY_CONTEXT,
-    SentrySpanAttribute,
+from sentry_sdk.profiler.continuous_profiler import (
+    try_autostart_continuous_profiler,
+    get_profiler_id,
+    try_profile_lifecycle_trace_start,
 )
+from sentry_sdk.profiler.transaction_profiler import Profile
 from sentry_sdk._types import TYPE_CHECKING
 
 if TYPE_CHECKING:
